Make the vsync startup warning hidable
Since 5b51eb3 the vsync warning is set to always show up during startup
if vsync is turned off.

This is problematic with Intel/AMD drivers on linux, when using TearFree
(on in radeon driver by default, not sure about Intel)

This commit adds a button for the user to be able to suppress this warning.
Closes #934